case import emulator : individual zip containers (#8)
* Create individual zip containers for SV, SSH, TP, EQ profiles (ignoring other profiles)
Deal with optional businessProcess part in filename for EQ profile
Deal with version part in filename, which should be 3 characters long


* Remove no more used second parameter of update_cgmes.sh script

+++ case-import-emulator/update_cgmes.sh
@@ -1,14 +1,12 @@
 #!/bin/bash
 
 cgmesPath=$1
-country=$2
-outputDir=$3
+outputDir=$2
 version=1
 
 # generate a new date for the CGMES
 caseDateIso=$(date -u +"%Y-%m-%dT%H:%M:%SZ")
-caseDate=$(date -d $caseDateIso +"%Y%m%d_%H%M")
-dayOfTheWeek=$(date -d $caseDateIso '+%u')
+caseDate=$(date -d $caseDateIso +"%Y%m%dT%H%MZ")
 
 # create a working directory
 tmpDir=$(mktemp -d)
@@ -25,19 +23,29 @@ unzip *.zip >> /dev/null
 caseName=${zipFile%%.zip}
 cd $caseName
 for xmlFile in $(ls *.xml); do
-    sed -i -e "s/<md:Model.scenarioTime>.*<\/md:Model.scenarioTime>/<md:Model.scenarioTime>$caseDateIso<\/md:Model.scenarioTime>/g" $xmlFile
-    sed -i -e "s/<md:Model.version>.*<\/md:Model.version>/<md:Model.version>$version<\/md:Model.version>/g" $xmlFile
+    # we only deal with SV, TP, SSH and EQ profiles
+    if [[ "$xmlFile" =~ ^.*_(SV|TP|EQ|SSH)_[0-9]+[.]xml$ ]] 
+    then
+      sed -i -e "s/<md:Model.scenarioTime>.*<\/md:Model.scenarioTime>/<md:Model.scenarioTime>$caseDateIso<\/md:Model.scenarioTime>/g" $xmlFile
+      sed -i -e "s/<md:Model.version>.*<\/md:Model.version>/<md:Model.version>$version<\/md:Model.version>/g" $xmlFile
+      zipFile=${xmlFile%%.xml}
+      newZipFile=${caseDate}_${zipFile#*Z_}
+      # for EQ profile, deal with optional businessProcess part in file name (two underscores after dateTime needed in that case)
+      if [[ "$newZipFile" =~ ^.*_EQ_[0-9]+.*$ && ! "$newZipFile" =~ ^.*Z__.*_EQ_.*$ ]]
+      then
+        newZipFile=${newZipFile%Z_*}Z__${newZipFile#*Z_}
+      fi
+      # add version padding with zeros (3 characters needed)
+      newZipFile=${newZipFile%*_[0-9]}_`printf "%03d" ${newZipFile##*_}` 
+      # rebuild the zip with update profile and using a new naming
+      zip -j $newZipFile $xmlFile >> /dev/null
+      # copy the new CGMES to output directory
+      rm -f $outputDir/${newZipFile}.zip
+      mv ${newZipFile}.zip $outputDir
+    fi
 done
 
-# rebuild the zip with update profiles and using a new naming
 cd ..
-rm $zipFile
-newZipFile=${caseDate}_FO${dayOfTheWeek}_${country}${version}
-zip -j $newZipFile $caseName/* >> /dev/null
-
-# copy the new CGMES to output directory
-rm -f $outputDir/${newZipFile}.zip
-mv ${newZipFile}.zip $outputDir
 
 # delete the working directory
 rm -rf $tmpDir
